Remove typescript from app
