+++ bench/heavy-npm-deps/app/page.js
@@ -1,0 +1,13 @@
+import { LodashComponent } from '../components/lodash'
+// import { MantineComponent } from "../components/mantine";
+// import { MermaidComponent } from "../components/mermaid";
+
+export default function Page() {
+  return (
+    <>
+      {/* <MantineComponent /> */}
+      {/* <MermaidComponent /> */}
+      <LodashComponent />
+    </>
+  )
+}
