--- bench/heavy-npm-deps/app/page.tsx
@@ -1,13 +1,0 @@
-import { LodashComponent } from '@/components/lodash'
-// import { MantineComponent } from "@/components/mantine";
-// import { MermaidComponent } from "@/components/mermaid";
-
-export default function Page() {
-  return (
-    <>
-      {/* <MantineComponent /> */}
-      {/* <MermaidComponent /> */}
-      <LodashComponent />
-    </>
-  )
-}
