+++ bench/heavy-npm-deps/components/lodash.js
@@ -1,12 +1,12 @@
-"use client";
-import * as Lodash from "lodash-es";
+'use client'
+import * as Lodash from 'lodash-es'
 
-console.log(Lodash);
+console.log(Lodash)
 
 export function LodashComponent() {
   return (
     <>
       <h1>Client Component</h1>
     </>
-  );
+  )
 }
