+++ bench/heavy-npm-deps/components/mantine.js
