+++ bench/heavy-npm-deps/components/mermaid.js
